+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#6237](https://github.com/influxdata/influxdb/issues/6237): Enable continuous integration testing on Windows platform via AppVeyor. Thanks @mvadu
 - [#6263](https://github.com/influxdata/influxdb/pull/6263): Reduce UDP Service allocation size.
 - [#6228](https://github.com/influxdata/influxdb/pull/6228): Support for multiple listeners for collectd and OpenTSDB inputs.
+- [#6292](https://github.com/influxdata/influxdb/issues/6292): Allow percentile to be used as a selector.
 
 ### Bugfixes